+++ Changelog.md
@@ -1,4 +1,5 @@
 - (***ADDB***) Fix Modify Alignment features causing saves to become unloadable.
+- (***ADDB***) Add Settings => "Fix save that broke because of alignment" for saves which got affected by the above bug. Just activate this and you should be able to load the save that was broken by that issue.
 - (***ADDB***) Add support for BlueprintClueStudy to Remote Companion Dialog (probably mostly noticeable when investigating clues).
 - (***ADDB***) Fix my save fix hotfix applying outside of saving routine too (should make no difference).